pyproject.toml
[project]
name = "projetogaragem-django"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Gabz047", email = "gabe1612mail@gmail.com"},
]
dependencies = [
    "django>=5.0.3",
    "djangorestframework>=3.15.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
